+++ client/src/components/Profile/Forms/FormDatePicker.jsx
@@ -1,7 +1,7 @@
-import React from 'react';
 import { Grid, InputLabel } from '@mui/material';
 import { DatePicker } from '@mui/x-date-pickers';
 import PropTypes from 'prop-types';
+import React from 'react';
 
 export default function FormDatePicker(props) {
   const {
