+++ client/src/components/Profile/Forms/FormTextFieldInput.jsx
@@ -1,8 +1,8 @@
-import React from 'react';
 import {
   Grid, InputAdornment, InputLabel, TextField,
 } from '@mui/material';
 import PropTypes from 'prop-types';
+import React from 'react';
 
 export default function FormTextFieldInput(props) {
   const {
