+++ client/src/pages/Profile.jsx
@@ -2,7 +2,7 @@ import { Typography } from '@mui/material';
 import React from 'react';
 
 import AvatarUpload from '../components/Profile/AvatarUpload';
-import Form from '../components/Profile/Forms/Form';
+import BasicInfoForm from '../components/Profile/Forms/BasicInfoForm';
 
 export default function Profile() {
   return (
@@ -11,7 +11,7 @@ export default function Profile() {
         Profile Page
       </Typography>
       <AvatarUpload />
-      <Form />
+      <BasicInfoForm />
     </>
   );
 }
